JSON Parser Test Suite (#232)
* feat: port tests from jsmnsol lib

The truffle tests were pretty minimal, but good to keep in sync since
this isn't a published package we're importing.
https://github.com/chrisdotn/jsmnSol/tree/master/test

* feat: add jsc JSONTestSuite

Since the existing test suite didn't catch any of the reported problems
with the JSON parsing, adding another larger suite that's used to
compare popular JSON parsers should!

Obviously because this is so large it can be trimmed, but the hope is to
find overlap between the feedback and the test-cases here.

* fix: update to latest era-test-node

This action is broken again, hopefully the latest works!

* fix: provide more debug for json tests

Tests connect to the files

* fix: parse an emoji instead of backslash

Skipping all the json format tests for now,
These are helpful if we want a fully compliant parser, but aren't
helpful in parsing the client data JSON.

* feat: add duplicate key parsing test

It doesn't return an error, because it doesn't actually record key
values

* fix: undo era-test-node version

it looks like this is fixed now

* Update test/JsmnSolLib.ts


* Update test/JsmnSolLib.ts


* Update test/JsmnSolLib.ts


* feat: add json string tests

Covers existing functionality

* fix: re-add removed return result

---------

+++ src/libraries/JsmnSolLib.sol
@@ -305,7 +305,7 @@ library JsmnSolLib {
   }
 
   // parseInt
-  function parseInt(string memory _a) internal pure returns (int) {
+  function parseInt(string memory _a) internal pure returns (int256) {
     return parseInt(_a, 0);
   }
 
@@ -342,7 +342,7 @@ library JsmnSolLib {
     }
   }
 
-  function strCompare(string memory _a, string memory _b) internal pure returns (int) {
+  function strCompare(string memory _a, string memory _b) internal pure returns (int256) {
     bytes memory a = bytes(_a);
     bytes memory b = bytes(_b);
     uint256 minLength = a.length;
